inout changes
type now 'input' or 'output' instead of 1 or 0. We might want
to change this to xcos type (IN_f etc) if we need to track what
type of port the inout is for, and not just the direction.
inout name now reflects port type and index which is the best label
we can do.