preprocess diagram
used to preprocess the diagram in various ways before updating.
The diagram is manipulated in various ways to aid in the generation
of HDL at a later stage.
At the moment dummy helper blocks are added after input ports
and before output ports. These blocks will have their port types
updated during update but will not be removed as the input and
output ports are. They can then be used later during HDL generation
to communicate this info to scripts.
Further work is planned to do something similar for inport and
outport blocks, and GOTOs and FROMs.