simpler ratel library
removed adder from library. Intention is to reuse standard
xcos blocks (perhaps have them in palettes in the ratel
section of the library). This will hopefully mean a minimum
of specialist blocks.

It may mean modification of xcos stuff but being closer linked
to, and being able to modify internals will hopefully mean a better
product and leverages the benefits of open source.